fixup! fixup! fixup! fixup! Fixes #21117: Moving create node plugin into Rudder
Fixes #21117: Moving create node plugin into Rudder
